chore: move generic rules into separate file
